LT report 接口

+++ takin-web-biz-service/src/main/java/io/shulie/takin/web/biz/pojo/output/report/ReportAppMapOut.java
@@ -26,6 +26,6 @@ public class ReportAppMapOut {
     private Double successRate[];
 
     @ApiModelProperty(value = "请求数数组")
-    private Double totalRequest[];
+    private Integer totalRequest[];
 
 }
